+++ src/io/tts.py
@@ -1,9 +1,9 @@
 import logging
-import re
 
 from RealtimeTTS import KokoroEngine, SystemEngine, TextToAudioStream
 
 from config import config
+from src.util import strip_markdown
 
 log = logging.getLogger(f"{config.name}.{__name__}")
 
@@ -60,34 +60,3 @@ def init_tts():
     TextToAudioStream(engine).feed([text]).play(log_synthesized_text=True)
 
     engine.shutdown()
-
-
-def strip_markdown(md_text: str):
-    # strip emojies
-    text = "".join(char for char in md_text if char.isprintable())
-    # Remove links
-    text = re.sub(r"\[([^\]]+)\]\([^)]+\)", r"\1", text)
-
-    # Remove bold/italic (**text**, *text*, __text__, _text_)
-    text = re.sub(r"(\*\*|__)(.*?)\1", r"\2", text)
-    text = re.sub(r"(\*|_)(.*?)\1", r"\2", text)
-
-    # Remove headings (#, ##, ###, etc.)
-    text = re.sub(r"^#+\s*", "", text, flags=re.MULTILINE)
-
-    # Remove inline code (`code`)
-    text = re.sub(r"`(.*?)`", r"\1", text)
-
-    # Remove images while ensuring no extra spaces remain
-    text = re.sub(r"!\[.*?\]\(.*?\)", "", text)
-    text = re.sub(
-        r"\s*!\[.*?\]\(.*?\)\s*", " ", text
-    )  # Extra safeguard for inline images
-
-    # Remove unordered list markers (-, *, +)
-    text = re.sub(r"^\s*[-*+]\s*", "", text, flags=re.MULTILINE)
-
-    # Remove ordered list markers (1., 2., 3.)
-    text = re.sub(r"^\s*\d+\.\s*", "", text, flags=re.MULTILINE)
-
-    return text
